chat-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.54.0
python-dotenv>=1.0.1
httpx>=0.27.0
